Added maxheap.c
Maxheap is a classical data structure, defined as a specialised full binary tree data structure where nodes are arranged based on a node value. The following attached code file is a minimal example of a maxheap. Please excuse the variable names as this code was written in the infancy of my C programming journey